+++ CHANGELOG.md
@@ -484,3 +484,7 @@
 * Equipment slots are now defined by a pseudo-extensible enum and stored as int
 * Removed unused Ascension III specific LevelUpModal
 * RPG skills and stats are now defined by a pseudo-extensible enum and stored as int
+* CharacterPanelController no longer breaks if an expected stat, still, etc has no value on the character
+* WeaponSkillType is no longer used (use SkillType instead) and is obsolete
+* All equipped armor items are now factored into player damage protection/threshold calculations
+* All equipped items will be considered for shields after ShieldGenerator and Body are tried